requirements-gpu.txt
numpy==1.23.5
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
matplotlib==3.7.3
scipy==1.11.3
cupy-cuda11x==12.0.0
